Fix $PATH resolution
Because idx has not been updated when adding a version but discovery was relying on it to fetch it from store back, this resulted in getting a version under wrong idx.

Problem has been solved by not getting the version unnecessarily from store but local variable. This made returning the idx from addVersion obsolete.